% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/davidClustering.R
\name{davidClustering}
\alias{davidClustering}
\title{Cluster rows of the matrix by the hierarhical fuzzy multi-linkage partitioning method proposed by DAVID}
\usage{
davidClustering(matrix, kappaThr = 0.35, initialGroupMembership = 3,
  multiLinkageThr = 0.5, removeRedundant = TRUE, debug = FALSE)
}
\arguments{
\item{matrix}{Input An adjacency matrix: the values in the cells are either 0 or 1.}

\item{kappaThr}{Numeric, the threshold of the Kappa statistic, which is used to select initial seeds. Default value: 0.35, as recommended by the authors of the original study based on their experiences.}

\item{initialGroupMembership}{Integer, the number of minimal members in initial groups. Default value: 3.}

\item{multiLinkageThr}{Numeric, the minimal linkage between two groups to be merged. Default value: 0.5.}

\item{removeRedundant}{Logical, whether redundant initial groups should be removed before clustering. Used for debugging. Setting as \code{TRUE} accelerates the iterative clustring process.}

\item{debug}{Logical, whether seed information is printed for debugging purposes.}
}
\description{
The function implements the Hierarhical fuzzy multi-linkage partitioning method used in the DAVID Bioinformatics tool.
}
\note{
The function has only been tested in a few anecdotal examples. Cautions and more systematic tests are required before it is applied to critical datasets.
}
\references{
\itemize{
\item{Huang *et al.* The DAVID Gene Functional Classification Tool: a novel biological module-centric algorithm to functionally analyze large gene lists. Genome Biology, 2007}
\item{Additional file of the manuscript available at \url{https://david.ncifcrf.gov/helps/2D_Introduction_files/additional_file_13.doc}}
}
}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
