% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{davidClustering_cpp}
\alias{davidClustering_cpp}
\title{Perform heuristic fuzzy multi-linkage partitioning of DAVID}
\usage{
davidClustering_cpp(kappaMatrix, kappaThr = 0.35,
  initialGroupMembership = 3L)
}
\description{
Perform heuristic fuzzy multi-linkage partitioning of DAVID
}
\examples{
testMat <- cbind(c(1,1,0,0,1,0), c(1,1,0,1,1,0), c(1,0,1,1,0,1))
testKappaMat <- rowKappa(testMat)
davidClustering_cpp(testKappaMat)

}
