% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{davidClustering_kappa}
\alias{davidClustering_kappa}
\title{Cluster rows of a Kappa-statistic-matrix by the hierarhical fuzzy multi-linkage partitioning method proposed by DAVID}
\usage{
davidClustering_kappa(kappaMatrix, kappaThr = 0.35,
  initialGroupMembership = 3L, multiLinkageThr = 0.5, mergeRule = 1L)
}
\arguments{
\item{kappaMatrix}{A numeric matrix of Kappa statistics, which is likely returned by \code{\link{rowKappa}} or \code{\link{colKappa}}}

\item{kappaThr}{Numeric, the threshold of the Kappa statistic, which is used to select initial seeds. Default value: 0.35, as recommended by the authors of the original study based on their experiences.}

\item{initialGroupMembership}{Integer, the number of minimal members in initial groups. Default value: 3.}

\item{multiLinkageThr}{Numeric, the minimal linkage between two groups to be merged. Default value: 0.5.}

\item{mergeRule}{Integer, how two seeds are merged. See below.

Merge rules:
\itemize{
\item{1: length of intersect divided by length of union no less than \code{multiLinkageThr}.}
\item{2: length of intersect divided by length of \emph{both} seeds no less than \code{multiLinkageThr}.}
\item{3: length of intersect divided by length of \emph{either} seeds no less than \code{multiLinkageThr}.}
\item{4: Geometric mean of length of intersect divided by length of \emph{both} seeds no less than \code{multiLinkageThr}.}
}}
}
\description{
The function implements the Hierarhical fuzzy multi-linkage partitioning method used in the DAVID Bioinformatics tool.
}
\note{
The function has only been tested in a few anecdotal examples. Cautions and more systematic tests are required before it is applied to critical datasets.
}
\examples{
synData <- matrix(c(rep(c(rep(1, 10), rep(0, 5)), 3),
rep(0, 4), rep(1, 7), rep(0,4),
rep(c(rep(0,5), rep(1,10)), 3),
rep(c(rep(0,3), 1), 4)[-16]), ncol=15, byrow=TRUE)
rownames(synData) <- sprintf("Gene \%s", letters[1:8])
colnames(synData) <- sprintf("t\%d", 1:15)
synKappaMat <- rowKappa(synData)
synKappaMat.round2 <- round(synKappaMat, 2)
davidClustering_kappa(synKappaMat.round2)

}
\references{
\itemize{
\item{Huang *et al.* The DAVID Gene Functional Classification Tool: a novel biological module-centric algorithm to functionally analyze large gene lists. Genome Biology, 2007}
\item{Additional file of the manuscript available at \url{https://david.ncifcrf.gov/helps/2D_Introduction_files/additional_file_13.doc}}
}
}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
