% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa.R
\name{kappaLA}
\alias{kappaLA}
\title{Calculate column-wise kappa statistics of a matrix, using linear algebra procedure}
\usage{
kappaLA(matrix)
}
\arguments{
\item{matrix}{a binary matrix of either 0 or one}
}
\value{
A matrix of size nxn if the input matrix is of size mxn (m is arbitrary)
}
\description{
Calculate column-wise kappa statistics of a matrix, using linear algebra procedure
}
\examples{
testMat <- cbind(c(1,1,0,0,1,0), c(1,1,0,1,1,0))
kappaLA(testMat)
}
\seealso{
\code{\link{kappaSimp}} to calculate the same statistic using a simpler routine
}
