% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa.R
\name{kappaSimp}
\alias{kappaSimp}
\title{Calculate column-wise kappa statistics of a matrix, using a simple procedure by going through the matrix and counting}
\usage{
kappaSimp(matrix, minOverlap = 0)
}
\arguments{
\item{matrix}{a binary matrix of either 0 or one}

\item{minOverlap}{Numeric/integer, the minimal overlap between two columns to be considered for further calculation}
}
\value{
A matrix of size nxn if the input matrix is of size mxn (m is arbitrary)
}
\description{
Calculate column-wise kappa statistics of a matrix, using a simple procedure by going through the matrix and counting
}
\seealso{
\code{\link{colKappa}} to calculate the same statistic using a linear algebra based routine

Other kappa functions: 
\code{\link{colKappa}()},
\code{\link{rowKappa}()}
}
\concept{kappa functions}
