% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa.R
\name{rowKappa}
\alias{rowKappa}
\title{Calculate row-wise kappa statistics of a matrix}
\usage{
rowKappa(matrix)
}
\arguments{
\item{matrix}{a binary matrix, containing values of either 0 or 1.}
}
\value{
A matrix of size \eqn{m \times m} if the input matrix is of size \eqn{m \times m}.
}
\description{
The function returns row-wise kappa statistics of a matrix, using a linear algebra procedure implemented in C++.
}
\examples{
testMat <- cbind(c(1,1,0,0,1,0), c(1,1,0,1,1,0))
rowKappa(testMat)

}
\seealso{
\code{\link{colKappa}} to calculate the statistic of columns

Other kappa functions: \code{\link{colKappa}},
  \code{\link{kappaSimp}}
}
