% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_from_density.R
\name{simulate_from_density}
\alias{simulate_from_density}
\title{Simulate from density
Compared with bootstrapping, the results do not reveal input values, and the
empirical distribution can be smoother. The function assumes that the data
is normally distributed.}
\usage{
simulate_from_density(vec, N = 1e+05)
}
\arguments{
\item{vec}{Numeric vector}

\item{N}{Integer, number of simulated instances}
}
\description{
Simulate from density
Compared with bootstrapping, the results do not reveal input values, and the
empirical distribution can be smoother. The function assumes that the data
is normally distributed.
}
\examples{
my_vec <- c(23, 27, 26, 24, 25)
simulate_from_density(my_vec, 10)
}
\author{
Iakov Davydov
}
