% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfidf.R
\name{tfidf}
\alias{tfidf}
\title{Calculate TF-IDF using a input matrix with terms in rows and documents in columns}
\usage{
tfidf(tdMat, tfVariant = c("raw", "binary", "frequency", "log",
  "doubleNorm0.5"), idfVariant = c("raw", "smooth", "probabilistic"),
  idfAddOne = TRUE)
}
\arguments{
\item{tdMat}{A term-document matrix, terms in rows, documents in columns, and counts as integers (or logical values) in cells}

\item{tfVariant}{Variant of term frequency. See details below.}

\item{idfVariant}{Variant of inverse document frequency. See details below.}

\item{idfAddOne}{Logical, whether one should be added to both numerator and denominator to calculate IDF. See details below.}
}
\description{
Calculate TF-IDF using a input matrix with terms in rows and documents in columns
}
\examples{
tiExample <- matrix(c(1,1,1,1,1,
1,1,0,0,0,
1,0,0,0,0,
0,1,0,0,0,
0,0,0,1,0,
1,0,1,0,1,
0,0,0,0,1), ncol=5, byrow=TRUE)
colnames(tiExample) <- sprintf("D\%d", 1:ncol(tiExample))
rownames(tiExample) <- sprintf("t\%d", 1:nrow(tiExample))
tiRes <- tfidf(tiExample)

}
