% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa.R
\name{kappaSimp}
\alias{kappaSimp}
\title{Calculate column-wise kappa statistics of a matrix, using a simple procedure by going through the matrix and counting}
\usage{
kappaSimp(matrix)
}
\arguments{
\item{matrix}{a binary matrix of either 0 or one}
}
\value{
A matrix of size nxn if the input matrix is of size mxn (m is arbitrary)

#' @family kappa functions
}
\description{
Calculate column-wise kappa statistics of a matrix, using a simple procedure by going through the matrix and counting
}
\examples{
testMat <- cbind(c(1,1,0,0,1,0), c(1,1,0,1,1,0))
kappaSimp(testMat)
}
\seealso{
\code{\link{kappaLA}} to calculate the same statistic using a linear algebra based routine
}

