% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{randmat}
\alias{randmat}
\title{Make a random matrix by sampling}
\usage{
randmat(x, n, N)
}
\arguments{
\item{N}{Integer, number of sampling repeats, or the column count of the matrix.}

\item{vec}{A numeric vector to be sampled from}

\item{size}{Integer, size of sampling, or the row count of the matrix.}
}
\value{
A random matrix of the dimension \eqn{size \times N}.
}
\description{
Generating a random matrix by sampling a numeric vector with replacement per column.
}
\details{
The function generates a matrix of the dimension \eqn{size \times N}. Each column of the matrix is a random sampling of the input vector.
  
Currently only sampling \emph{with replacement} is supported.
}
\examples{
testVec <- 1:10
randmat(testVec, size=5L, N=10L)
}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
